Simplify the tests.

+++ testing/pod-tests/master/pod-destroy.sh
@@ -8,10 +8,14 @@ while true; do
     sleep 1
     continue
   fi
+  kubectl -n minions get pods 2>/dev/null | grep minion | grep Pend
+  if [ $? -eq 0 ]; then
+    echo "Got pending item, see what's happening"
+    exit 1
+  fi
   echo "$i times of delete"
   ((i++))
   kubectl -n minions delete pod $pod
   #sleep 5
 done
 echo "over $i times"
-
